fix: bugs container and rename files
