+++ src/css/base/container.css
@@ -1,27 +1,19 @@
 .container {
-    min-width: 320px;
+    max-width: 375px;
     padding: 0 20px;
     margin: 0 auto;
 }
 
-@media screen and (min-width: 480px) {
-    .container {
-        min-width: 480px;
-        margin-left: auto;
-        margin-right: auto;
-    }
-}
-
 @media screen and (min-width: 768px) {
     .container {
-        min-width: 768px;
+        max-width: 768px;
         padding: 0 32px;
         overflow: hidden;
     }
 }
 
-@media screen and (min-width: 1280px) {
+@media screen and (min-width: 1440px) {
     .container {
-        width: 1280px;
+        max-width: 1440px;
     }
 }
