+++ src/css/components/header.css
@@ -4,7 +4,7 @@
     align-items: center;
     margin-top: 24px;
 
-    @media screen and (min-width: 1280px) {
+    @media screen and (min-width: 1440px) {
             margin-top: 28px;
         }
 }
@@ -27,13 +27,10 @@
         flex-shrink: 0;
         border-radius: 60px;
         background: var(--main-color-light-theme, #242424);
-    }
-    
-    @media screen and (min-width: 767px) and (max-width: 1279px) {
-        margin-right: calc(80px + ((100vw - 767px) / (1279 - 767) * (400 - 80)));
+        margin-right: 80px;
     }
 
-    @media screen and (min-width: 1280px) {
+    @media screen and (min-width: 1440px) {
         width: 265px;
         height: 52px;
         margin-right: 400px;
