+++ src/css/styles.css
@@ -13,12 +13,10 @@
 @import url('../css/components/exercises.css');
 @import url('../css/components/hero.css');
 @import url('../css/components/quote.css');
+@import url('../css/components/header.css');
 
 /*-----layouts-----*/
-@import url('./layouts/header.css');
 
-/*-----layouts-----*/
-@import url('./layouts/header.css');
 
 /*-----variables-----*/
 @import url('./utils/variables.css');
