+++ src/js/pageSwitcher.js
