+++ src/main.js
@@ -1,6 +1,6 @@
-import './js/page-switcher';
-import './js/mobile-menu';
+import './js/pageSwitcher';
+import './js/mobileMenu';
 import './js/modal';
-import './js/modal-rating';
+import './js/modalRating';
 import './js/exercises';
 import './js/quote';
